feat: adding code for logging with google
