+++ app.json
@@ -5,6 +5,7 @@
     "version": "1.0.0",
     "orientation": "portrait",
     "userInterfaceStyle": "light",
+    "scheme":"com.fifiuba.app",
     "splash": {
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
