+++ src/components/login-form.js
@@ -1,8 +1,30 @@
-import {React, useState} from 'react';
+import {useState} from 'react';
 import {View, StyleSheet, Text} from 'react-native';
 import {TextInput, Button, Colors} from 'react-native-paper';
 import {useForm, Controller} from 'react-hook-form';
 import login from '../services/login';
+
+import * as React from 'react';
+import * as WebBrowser from 'expo-web-browser';
+import * as Google from 'expo-auth-session/providers/google';
+import { initializeApp } from 'firebase/app';
+import { getAuth, GoogleAuthProvider, signInWithCredential } from 'firebase/auth';
+
+const firebaseConfig = {
+  apiKey: "AIzaSyD_fAr5j7fbdhezakqnBmSLW2xEZ9Uki2U",
+  authDomain: "user-service-9def8.firebaseapp.com",
+  databaseURL: "https://user-service-9def8-default-rtdb.firebaseio.com",
+  projectId: "user-service-9def8",
+  storageBucket: "user-service-9def8.appspot.com",
+  messagingSenderId: "627165168741",
+  appId: "1:627165168741:web:23447b751047a7ca24b28c",
+  measurementId: "G-YL8STET37V"
+};
+
+// Initialize Firebase
+initializeApp({firebaseConfig});
+
+WebBrowser.maybeCompleteAuthSession();
 
 const LoginForm = (props) => {
   const [hidePassword, sethidePassword] = useState(true);
@@ -18,6 +40,22 @@ const LoginForm = (props) => {
     login(data);
     props.onLogin(true);
   }
+
+  const WEB_CLIENT_ID = '627165168741-aqol4difl2m8tma737ipfq6pbn4jahsi.apps.googleusercontent.com';
+  const [_request, response, promptAsync] = Google.useIdTokenAuthRequest(
+    {
+      clientId: WEB_CLIENT_ID,
+    },
+  );
+
+  React.useEffect(() => {
+    if (response?.type === 'success') {
+      const { id_token } = response.params;
+      const auth = getAuth();
+      const credential = GoogleAuthProvider.credential(id_token);
+      signInWithCredential(auth, credential);
+    }
+  }, [response]);
 
   return (
     <View style={loginStyle.container}>
@@ -101,6 +139,9 @@ const LoginForm = (props) => {
           style={loginStyle.buttonRedes}
           color={Colors.red800}
           mode="contained"
+          onPress={() => {
+          promptAsync();
+        }}
         >
           <Text style={{fontSize: 18}}>Google</Text>
         </Button>
